fixed unittest
