+++ gbpservice/nfp/pecan/api/config.py
@@ -64,11 +64,5 @@ cloud_services = [
      'topic': 'configurator',
      'reporting_interval': '10',  # in seconds
      'apis': ['CONFIGURATION']
-     },
-
-    {'service_name': 'visibility',
-     'topic': 'visibility',
-     'reporting_interval': '10',  # in seconds
-     'apis': ['VISIBILITY']
-     },
+     }
 ]
